requirements.txt
Django==1.5.1
MySQL-python==1.2.4
django-keyedcache==1.4-6
django-livesettings==1.4-12
jsonrpclib==0.1.3
pyzmq==14.0.1
wsgiref==0.1.2
git+https://github.com/orokusaki/django-jsonrpc-2-0.git
git+http://bluewallvirtual.com/git/slipstream_py.git
git+https://github.com/dmgctrl/django-ztask.git
django-crispy-forms
django-bidi-utils
South
https://pypi.python.org/packages/source/d/django-keyedcache/django-keyedcache-1.5.1.tar.gz
#pyzmq
#-e git+git@github.com:dmgctrl/django-ztask.git#egg=django_ztask
django-oscar==0.7
django-oscar-paypal==0.9
disqus-python
django-social-auth
django-parsley
